[CORE][API] Adjust ISarosSessionManager to reference points
The ISarosSessionManager was adjusted to reference points and
works only with reference points and its resources now. So the
IReferencePointManager can not be filled with reference points and projects.
So IReferencePointManager will be filled in the CollaborationUtils class,
before the reference points and the ressources will be added to the
saros session.