[API] Adjust EditorManager to ReferencePoints
The EditorManager now works with reference points instead of core projects (especially saves the local editors
of all shared files, that belongs to given reference point).

Addionally some components were prepared for the next commits: The next commits handles the adjustment of
the SPath. The SPath will get the ReferencePointManager as class variable. More details come in the nexts
 commits.