[INTERNAL] Added IReferencePoint and its implementation
The Saros filesystem gets a new component in its filesystem:
IReferencePoint. The first step of the refactoring that the
reference point of each files/folders points on the location
of the projects. So the IResource interface gets a new method
called getReferencePoint() and returns the reference point.
The IReferencePoint interface is empty. Its implementation
contains an IPath object from the location of the referenced
resource (project at first) and an IReferencePoint object is
comparable over the IPath object.