[API][CORE] SharedProjectMapper: Using IReferencePoint
SharedProjectMapper/-test:
This class now with reference points with resources now.

IReferencePointManager/ReferencePointManager:
The IReferencePointManager maps an IReferencePoint to an core.IProject.
Because not all classes will not adapted at the same time, the IReferencePointManager
is needed for classes, who still works core.IProjects. The ReferencePointManager
is an implementation of IReferencePointManager.

SarosSession:
After the SharedProjectManager works with reference points, the SarosSession needs
the ÍReferencePointManager for getting IProjects given by IReferencePoints.
The IReferencePointManager will instantiated in SarosSession and kept in
the core session container. The IReferencePointManager will needed only if a session is
running, so that the reason, that the IReferencePointManager is in the core session
container. Currently the SarosSession is responsible for putting the pairs IReferencePoints
and core.IProjects, which should to be shared.