[CORE][API] Adjust AbstractOutgoingProjectNegotiation to referencePoints
The AbstractOutgoingProjectNegotiation now works with reference points.
It still takes a list of projects as parameter but a list of reference
points will created. The function createProjectNegotiationDataList takes
the list of reference points instead of a list of projects. The list
of projects as global variable will removed in a later commit