[CORE][API] Adjust IncomingProjectNegotiation to ReferencePoints
AbstractIncomingProjectNegotiation:
The run method, for executing the file system changes (like project
structure synchronization), now takes a mapping of reference point ids to
reference point. Also there is the referencePointManager needed, for example
determining the folder given by a path in the project structure synchronization.

ArchiveIncomingProjectNegotiation:
Because of the adjusting the AbtractIncomingProjectNegotiation to reference
point, like the actual transfer, the implementation of the overwritten
transfer method has to handle the reference point instead the core.IProject.

InstantIncomingProjectNegotiation:
See ArchiveIncomingProjectNegotiation

For the IDE parts:
For running the project negotiation, the mapping of project/reference point id
to core.IProject had to transform to reference point id to IReferencePoint