[CORE][API] ISarosSessionManager: Add startSession(Map<...>, IReferencePointManager)
The ISarosSessionManager was extended by
startSession(Map<IReferencePoint, List<IResource>>, IReferencePointManager). The
IReferencePointManager is needed, because it has to be filled before the session starts.
So the filled IReferencePointManager has to hand over to the saros manager and the
saros manager has to hand over the IReferencePointManager to the saros session.
So it was needed to extend the CTOR of SarosSession, which additionally takes the
IReferencePointManager.