[RELEASE-FIX][I] Add workaround to replace old project object
THIS IS A RELEASE FIX! DO NOT MERGE ONTO MASTER!

Makes the IntellijProjectLifecycle singleton.

Introduces a wrapper for the held project object as it is changeable
during the plugin lifecycle.

Adds a workaround to replace the old project object when a new project
is opened and the currently held project object is disposed. This is
needed as the internal Saros logic otherwise tries to work on the old
project object which is no longer valid, leading to an assertion error.

This is only a very rudimentary workaround as it still leaves users in a
headless state if they open two projects and then close the one opened
first. This state is not handled gracefully by Saros (it will cause
assertion errors shown to the user in the IDE when trying to start a
session). The headless state is resolved once the user opens another
project which will then be used for Saros.