Add merge and test to JGitFacade
After fetching results, it's necessary to merge them to change the current HEAD.
The feature should offer an automatic merge if the current HEAD is an ancestor (fast-forward-merge) respecting possible errors/leaks while handling the data inside a Saros session.
The introduced method is the first step to achieve this.