[INTERNAL][I] Adjust NotificationPanel to use the shared project
Adds the method UIProjectUtils.getSharedProject() which returns the
project object contained in the current session context if present. This
method is used to determine the project object to use in Notification
panel.

With this logic, notifications created through the NotificationPanel
will be displayed in all projects that are part of the session or all
open projects if there is no running session.